docs: break up the messages.md
