+++ messages/package1_version_create.md
@@ -1,0 +1,8 @@
+# package1VersionCreateCommandUploadFailureDefault
+
+Package version creation failed with unknown error
+
+# package1VersionCreateCommandUploadFailure
+
+Package upload failed.
+%s
