+++ messages/package_ancestry.md
@@ -1,0 +1,19 @@
+# idOrAliasNotFound
+
+Can't find package with id or alias: %s
+
+# unlockedPackageError
+
+Can’t display package ancestry. Package ancestry is available only for second-generation managed packages. Retry this command and specify a second-generation managed package or package version.
+
+# noVersionsError
+
+Can’t display package ancestry. The specified package has no associated released package versions. Retry this command after you create and promote at least one package version.
+
+# invalidId
+
+Can’t display the ancestry tree for %s. Specify a valid package ID (starts with 0Ho) or package version ID (starts with 04t), and try creating the ancestry tree again.
+
+# versionNotFound
+
+Can’t display the ancestry tree for %s. Verify the package version number (starts with 04t) or the package version alias listed in the sfdx-project.json file, and try creating the ancestry tree again.
