+++ messages/package_uninstall.md
@@ -1,0 +1,3 @@
+# uninstallErrorAction
+
+Verify installed package ID and resolve errors, then try again.
