+++ messages/package_version_create.md
@@ -62,14 +62,10 @@ Can’t create package version. Specifying an ancestor is available only for sec
 
 versionName is blank in sfdx-project.json, so it will be set to this default value based on the versionNumber: %s
 
-# releaseNotesUrl
-
-release notes URL
-
-# postInstallUrl
-
-post-install URL
-
 # malformedUrl
 
 The %s value "%s" from the command line or sfdx-project.json is not in the correct format for a URL. It must be a valid URL in the format "http://salesforce.com". More information: https://nodejs.org/api/url.html#url_url_strings_and_url_objects
+
+# invalidDaysNumber
+
+Provide a valid positive number for %s.
