+++ messages/version_number.md
@@ -1,0 +1,15 @@
+# errorInvalidBuildNumberToken
+
+The provided VersionNumber '%s' is invalid. Build number token must be a number or one of these tokens '%s'.
+
+# errorMissingVersionNumber
+
+The VersionNumber property must be specified.
+
+# errorInvalidMajorMinorPatchNumber
+
+VersionNumber parts major, minor or patch must be a number but the value found is [%s].
+
+# errorInvalidVersionNumber
+
+VersionNumber must be in the format major.minor.patch.build but the value found is [%s].
